+++ example/playback.py
@@ -1,41 +1,9 @@
 from argparse import ArgumentParser
-from typing import Optional, Tuple
 
 import cv2
-import numpy as np
+from helpers import colorize, convert_to_bgra_if_required
 
-from pyk4a import ImageFormat, PyK4APlayback
-
-
-def colorize(
-    image: np.ndarray,
-    clipping_range: Tuple[Optional[int], Optional[int]] = (None, None),
-    colormap: int = cv2.COLORMAP_HSV,
-) -> np.ndarray:
-    if clipping_range[0] or clipping_range[1]:
-        img = image.clip(clipping_range[0], clipping_range[1])
-    else:
-        img = image.copy()
-    img = cv2.normalize(img, None, 0, 255, cv2.NORM_MINMAX, dtype=cv2.CV_8U)
-    img = cv2.applyColorMap(img, colormap)
-    return img
-
-
-def convert_to_bgra_if_required(color_format, color_image):
-    # examples for all possible pyk4a.ColorFormats
-    if color_format == ImageFormat.COLOR_MJPG:
-        color_image = cv2.imdecode(color_image, cv2.IMREAD_COLOR)
-    elif color_format == ImageFormat.COLOR_NV12:
-        color_image = cv2.cvtColor(color_image, cv2.COLOR_YUV2BGRA_NV12)
-        # this also works and it explains how the COLOR_NV12 color color_format is stored in memory
-        # h, w = color_image.shape[0:2]
-        # h = h // 3 * 2
-        # luminance = color_image[:h]
-        # chroma = color_image[h:, :w//2]
-        # color_image = cv2.cvtColorTwoPlane(luminance, chroma, cv2.COLOR_YUV2BGRA_NV12)
-    elif color_format == ImageFormat.COLOR_YUY2:
-        color_image = cv2.cvtColor(color_image, cv2.COLOR_YUV2BGRA_YUY2)
-    return color_image
+from pyk4a import PyK4APlayback
 
 
 def info(playback: PyK4APlayback):
