+++ example/viewer_depth.py
@@ -1,5 +1,6 @@
 import cv2
 import numpy as np
+from helpers import colorize
 
 import pyk4a
 from pyk4a import Config, PyK4A
@@ -24,13 +25,7 @@ def main():
     while True:
         capture = k4a.get_capture()
         if np.any(capture.depth):
-            # Clip 5000 mm(5 meters)
-            clipped_depth = capture.depth.clip(None, 5000)
-            # normalize and convert to 8bit
-            normalized_depth = cv2.normalize(clipped_depth, None, 0, 255, cv2.NORM_MINMAX, dtype=cv2.CV_8U)
-            # coloring image by choosed color map
-            colored_depth = cv2.applyColorMap(normalized_depth, cv2.COLORMAP_HSV)
-            cv2.imshow("k4a", colored_depth)
+            cv2.imshow("k4a", colorize(capture.depth, (None, 5000), cv2.COLORMAP_HSV))
             key = cv2.waitKey(10)
             if key != -1:
                 cv2.destroyAllWindows()
