+++ example/viewer_transformation.py
@@ -1,24 +1,8 @@
-from typing import Optional, Tuple
-
 import cv2
-import numpy as np
+from helpers import colorize
 
 import pyk4a
 from pyk4a import Config, PyK4A
-
-
-def colorize(
-    image: np.ndarray,
-    clipping_range: Tuple[Optional[int], Optional[int]] = (None, None),
-    colormap: int = cv2.COLORMAP_HSV,
-) -> np.ndarray:
-    if clipping_range[0] or clipping_range[1]:
-        img = image.clip(clipping_range[0], clipping_range[1])
-    else:
-        img = image.copy()
-    img = cv2.normalize(img, None, 0, 255, cv2.NORM_MINMAX, dtype=cv2.CV_8U)
-    img = cv2.applyColorMap(img, colormap)
-    return img
 
 
 def main():
